+++ src/Contract/LimitResolverInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JacyImp\ApiPlatformRateLimiter\Contract;
+
+interface LimitResolverInterface
+{
+    public function resolve(): int;
+}